requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-talisman==1.1.0
google-generativeai==0.8.3
gunicorn==22.0.0
pytest==8.2.0
python-dotenv==1.0.1

